+++ packages/utils/test/parser/schemaParser.test.ts
@@ -4,8 +4,10 @@ import {
   RECURSIVE_REF,
   RECURSIVE_REF_ALLOF,
   SCHEMA_DEPENDENCIES,
+  SCHEMA_WITH_ALLOF_CANNOT_MERGE,
   SCHEMA_AND_ONEOF_REF_DEPENDENCIES,
   SCHEMA_AND_REQUIRED_DEPENDENCIES,
+  SCHEMA_WITH_ARRAY_CONDITION,
   SCHEMA_WITH_ONEOF_NESTED_DEPENDENCIES,
   SCHEMA_WITH_SINGLE_CONDITION,
   SCHEMA_WITH_MULTIPLE_CONDITIONS,
@@ -56,6 +58,14 @@ describe('schemaParser()', () => {
   });
   it('parses superSchema properly', () => {
     const schemaMap = schemaParser(SUPER_SCHEMA);
+    expect(schemaMap).toMatchSnapshot();
+  });
+  it('parse schema with array condition', () => {
+    const schemaMap = schemaParser(SCHEMA_WITH_ARRAY_CONDITION);
+    expect(schemaMap).toMatchSnapshot();
+  });
+  it('parse schema with allof not able to merge', () => {
+    const schemaMap = schemaParser(SCHEMA_WITH_ALLOF_CANNOT_MERGE);
     expect(schemaMap).toMatchSnapshot();
   });
 });